Makes a bunch of things material-respecting
Rods, lattices, grilles and windows are now material based. Different materials affect looks and health mostly.
Can make rods of arbitrary materials now, as long as they're hard enough (50 hardness or more).
Can reinforce sheet stacks with rods of other materials. Currently only does something when building windows from resulting stacks. Can get reinforcement material back with a welder.
90% changes are just repathing rods to material stack

+++ code/game/objects/structures/lattice.dm
@@ -8,33 +8,38 @@
 	w_class = ITEM_SIZE_NORMAL
 	plane = ABOVE_PLATING_PLANE
 	layer = LATTICE_LAYER
+	color = COLOR_STEEL
+	var/material/material
+	var/init_material = MATERIAL_STEEL
 	//	obj_flags = OBJ_FLAG_CONDUCTIBLE
 
-/obj/structure/lattice/Initialize()
+/obj/structure/lattice/get_material()
+	return material
+
+/obj/structure/lattice/Initialize(mapload, var/new_material)
 	. = ..()
-///// Z-Level Stuff
 	if(!(istype(src.loc, /turf/space) || istype(src.loc, /turf/simulated/open)))
-///// Z-Level Stuff
 		return INITIALIZE_HINT_QDEL
+	if(!new_material)
+		new_material = init_material
+	material = SSmaterials.get_material_by_name(new_material)
+	if(!istype(material))
+		return INITIALIZE_HINT_QDEL
+
+	name = "[material.display_name] lattice"
+	desc = "A lightweight support [material.display_name] lattice."
+	color =  material.icon_colour
+
 	for(var/obj/structure/lattice/LAT in loc)
 		if(LAT != src)
 			crash_with("Found multiple lattices at '[log_info_line(loc)]'")
 			qdel(LAT)
 	icon = 'icons/obj/smoothlattice.dmi'
 	icon_state = "latticeblank"
-	updateOverlays()
-	for (var/dir in GLOB.cardinal)
-		var/obj/structure/lattice/L
-		if(locate(/obj/structure/lattice, get_step(src, dir)))
-			L = locate(/obj/structure/lattice, get_step(src, dir))
-			L.updateOverlays()
+	update_icon()
 
 /obj/structure/lattice/Destroy()
-	for (var/dir in GLOB.cardinal)
-		var/obj/structure/lattice/L
-		if(locate(/obj/structure/lattice, get_step(src, dir)))
-			L = locate(/obj/structure/lattice, get_step(src, dir))
-			L.updateOverlays(src.loc)
+	update_neighbors(src)
 	. = ..()
 
 /obj/structure/lattice/ex_act(severity)
@@ -60,10 +65,10 @@
 		var/obj/item/weapon/weldingtool/WT = C
 		if(WT.remove_fuel(0, user))
 			to_chat(user, "<span class='notice'>Slicing lattice joints ...</span>")
-		new /obj/item/stack/rods(loc)
+		new /obj/item/stack/material/rods(loc, 1, material.name)
 		qdel(src)
-	if (istype(C, /obj/item/stack/rods))
-		var/obj/item/stack/rods/R = C
+	if (istype(C, /obj/item/stack/material/rods))
+		var/obj/item/stack/material/rods/R = C
 		if(R.use(2))
 			src.alpha = 0
 			playsound(src, 'sound/weapons/Genhit.ogg', 50, 1)
@@ -75,22 +80,23 @@
 			return
 	return
 
-/obj/structure/lattice/proc/updateOverlays()
-	//if(!(istype(src.loc, /turf/space)))
-	//	qdel(src)
-	spawn(1)
-		overlays = list()
+/obj/structure/lattice/proc/update_neighbors(var/exclude)
+	for (var/dir in GLOB.cardinal)
+		var/obj/structure/lattice/L = locate(/obj/structure/lattice, get_step(src, dir))
+		if(L)
+			L.update_icon(exclude)
 
-		var/dir_sum = 0
+/obj/structure/lattice/update_icon(var/exclude)
+	var/dir_sum = 0
+	for (var/direction in GLOB.cardinal)
+		var/turf/T = get_step(src, direction)
+		var/obj/structure/S = locate(/obj/structure/lattice) in T
+		if(S && S != exclude)
+			dir_sum |= direction
+		else if(locate(/obj/structure/catwalk, T))
+			dir_sum |= direction
+		else
+			if(!(istype(get_step(src, direction), /turf/space)) && !(istype(get_step(src, direction), /turf/simulated/open)))
+				dir_sum |= direction
 
-		var/turf/T
-		for (var/direction in GLOB.cardinal)
-			T = get_step(src, direction)
-			if(locate(/obj/structure/lattice, T) || locate(/obj/structure/catwalk, T))
-				dir_sum += direction
-			else
-				if(!(istype(get_step(src, direction), /turf/space)) && !(istype(get_step(src, direction), /turf/simulated/open)))
-					dir_sum += direction
-
-		icon_state = "lattice[dir_sum]"
-		return
+	icon_state = "lattice[dir_sum]"
